Updated to meet the wp coding style more appropriately
