--- classes/Stock.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace WPC2O;
-
-class Stock
-{
-    public function StockSyncView(): string
-    {
-        return '<h2>Stock sync</h2>';
-    }
-}
