+++ includes/CONSTANTS.php
@@ -1,13 +1,13 @@
 <?php
 
-define("WP_C2O_VERSION", '1.0.0');
+define("WP_C2O_VERSION", "1.0.0");
 
 // API option values
-define("WPC2O_API_KEY", "wpc2o_api_key");
-define("WPC2O_API_ENDPOINT", "wpc2o_endpoint");
-define("WPC2O_API_STOCK_ENDPOINT", "wpc2o_stock_endpoint");
-define("WPC2O_API_STORE_MANAGER_EMAIL", "wpc2o_manager_email");
+define("WPC2O_API_KEY", "_wpc2o_api_key");
+define("WPC2O_API_ENDPOINT", "_wpc2o_endpoint");
+define("WPC2O_API_STOCK_ENDPOINT", "_wpc2o_stock_endpoint");
+define("WPC2O_API_STORE_MANAGER_EMAIL", "_wpc2o_manager_email");
 
 // Plugin option values
-define("WPC2O_DELIVERY_OPTION", 'wpc2o_delivery_options_selection');
-define("WPC2O_LOGO", 'wpc2o_logo_selection');
+define("WPC2O_DELIVERY_OPTION", "_wpc2o_delivery_options_selection");
+define("WPC2O_LOGO", "_wpc2o_logo_selection");
