+++ includes/wpc2o_options_logo.php
@@ -1,0 +1,6 @@
+<?php
+
+function wpc2o_get_logo_view(): string
+{
+    return '';
+}
