+++ includes/wpc2o_options_stock.php
@@ -1,0 +1,6 @@
+<?php
+
+function wpc2o_get_stock_view(): string
+{
+    return '<h2>Stock sync</h2>';
+}
